+++ src/Vector/Core/Exception/ElementNotFoundException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Vector\Core\Exception;
+
+/**
+ * Class ElementNotFoundException
+ * @package Vector\Control\Exception
+ */
+class ElementNotFoundException extends \Exception
+{
+    
+}