+++ changelog.md
@@ -4,7 +4,7 @@
 - php 7+ requirement
 - update Module names
   - Lib\Object -> Lib\Objects
-  - Lib\ArrayList -> Lib\Arrays
-- strict array params in Lib\Arrays (formerly Lib\ArrayList)
+  - Lib\Arrays -> Lib\Arrays
+- strict array params in Lib\Arrays (formerly Lib\Arrays)
 - Maybe/Either/Just
 - Pattern Matching in Modules